Hack: add a special calling convention to force indirect return
In render-dump we use `extern "C"` functions to implement and call C++
functions defined in the game. This works mostly fine, even though C and
C++ calling conventions are not exactly the same.

Where the ABI breaks is when C++ functions start returning objects that
have destructor. C++ forces them to be returned indirectly (via a
pointer passed by the calling in X8 register). Rust, however, has
different rules: it only does this if the object size is >128 bits.

To bridge this small gap we do a hack: patch the rust compiler to
support a new "force-indirect-return" calling convention that always
returns the result via the X8 register (corresponding to an argument
having `sret` attribute on the LLVM IR level).

This commit does exactly that: adds a "force-indirect-return" calling
convention and makes the return value on aarch64 always being passed
indirectly.